use base activity
